Refactoring components started
