--- app/components/AirgrabForm/messages.js
@@ -1,13 +1,0 @@
-/*
- * CreateAccountForm Messages
- *
- * This contains all the text for the CreateAccountForm component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.CreateAccountForm.header',
-    defaultMessage: 'This is the CreateAccountForm component !',
-  },
-});
