--- app/components/BidNameForm/messages.js
@@ -1,13 +1,0 @@
-/*
- * BidNameForm Messages
- *
- * This contains all the text for the BidNameForm component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.BidNameForm.header',
-    defaultMessage: 'This is the BidNameForm component !',
-  },
-});
