+++ app/components/Disclaimer/index.js
@@ -1,0 +1,12 @@
+//Standard disclaimer for executing EOS transactions
+import React from 'react';
+
+const Disclaimer = () => (
+  <p>
+    By executing this action you are agreeing to the EOS constitution
+    and this action{"'"}s associated ricardian contract. The ricardian contract
+    may be viewed in the Scatter approval prompt.
+  </p>
+);
+
+export default Disclaimer;
