--- app/components/DonateForm/messages.js
@@ -1,13 +1,0 @@
-/*
- * TransferForm Messages
- *
- * This contains all the text for the TransferForm component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.TransferForm.header',
-    defaultMessage: 'This is the TransferForm component !',
-  },
-});
