+++ app/components/Information/BlockOneLetter.js
@@ -1,0 +1,35 @@
+import React from 'react';
+
+const BlockOneLetter = () => {
+  return (
+    <div>
+      <h5>Dear GenerEOS supporters,</h5>
+      <p>
+        Block One recently announced that that they will be voting for Block Producers that share the core values
+        necessary to maximize the integrity and potential of the EOS public blockchain network.
+      </p>
+      <p>
+        They have provided an email for BP&apos;s and token holders to campaign for the Block Producers they
+        believe in.
+      </p>
+      <p>
+        We would love if you could rally behind us showing your support by sending a passionate email to Block One
+        explaining why they should vote for GenerEOS.
+      </p>
+      <p>
+        Send your email to - <a href="mailto:bp@eos.io?Subject=I%20Support%20GenerEOS">bp@eos.io</a>
+      </p>
+      <p> 
+        This support would mean the world to us and help us keep creating great tools and contributions for the
+        community.
+      </p>
+      <p>
+        <strong>&hearts;,</strong>
+        <br />
+        Team GenerEOS
+      </p>
+    </div>
+  )
+}
+
+export default BlockOneLetter;
