+++ app/components/Information/Donate.js
@@ -1,0 +1,34 @@
+import React from 'react';
+
+const Donate = () => {
+  return (
+    <div>
+      <h5>Dear EOSToolkit supporters,</h5>
+      <p>
+        The <strong>EOSToolkit</strong> was made out of love prior to the launch of the EOS mainnet thanks to long
+        hours late into the night and a staunch belief in making EOS accessible to everyone.
+      </p>
+      <p>
+        Thanks to your support and votes <strong>Team GenerEOS</strong> now receives block rewards and this helps
+        fund further development of the eostoolkit and other EOS projects.
+      </p>
+      <p>
+        However, the goal of GenerEOS is to give back to the community.
+      </p>
+      <p>
+        By donating you will be supporting development of the <strong>EOSToolkit</strong> directly, while allowing
+        GenerEOS to give even more back to those who need it most.
+      </p>
+      <p>
+        We hope you love the <strong>EOSToolkit</strong> as much as I{"'"}ve loved making it.
+      </p>
+      <p>
+        <strong>&hearts;,</strong>
+        <br />
+        Nathan Rempel, Team GenerEOS
+      </p>
+    </div>
+  )
+}
+
+export default Donate;
