+++ app/components/Information/PoormanInfo.js
@@ -1,0 +1,21 @@
+import React from 'react';
+//Poorman token info
+const PoormanInfo = () => {
+  return (
+    <div>
+      <h5>The POORMANTOKEN is an EOS token that includes signups and burning to make airdrops cheaper</h5>
+      <p>The intention is to allow EOS accounts to Airgrab the token (i.e. opt-in or signup)</p>
+      <p>You can participate in this test by using this Airgrab form.</p>
+      <p>The July 11 drop was successful! More will be announced later, keep registering so you can catch up!</p>
+      <p>There may be future drops to test different airdrop and signup criteria</p>
+      <p>
+        To read more about this token contract check out the{' '}
+        <a href="https://github.com/generEOS/poorman.token" target="new">
+          poorman.token Github
+        </a>
+      </p>
+    </div>
+  )
+}
+
+export default PoormanInfo;
