+++ app/components/Pages/GovernancePage/Loadable.js
@@ -1,5 +1,5 @@
 /**
- * Asynchronously loads the component for HomePage
+ * Asynchronously loads the component
  */
 import Loadable from 'react-loadable';
