+++ app/components/Tool/Tool.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import GridContainer from 'components/Grid/GridContainer';
+
+const Tool = props => {
+  return (
+    <GridContainer>
+      {props.children}
+    </GridContainer>
+  );
+};
+
+export default Tool;
