+++ app/containers/Airgrab/index.js
@@ -7,7 +7,7 @@
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import { compose } from 'redux';
-import AirgrabForm from 'components/AirgrabForm';
+import AirgrabForm from 'components/Features/AirgrabForm';
 import { makeSelectEosAccount } from 'containers/Scatter/selectors';
 
 import injectSaga from 'utils/injectSaga';
