--- app/containers/Governance/messages.js
@@ -1,13 +1,0 @@
-/*
- * HomePage Messages
- *
- * This contains all the text for the HomePage component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  header: {
-    id: 'app.components.HomePage.header',
-    defaultMessage: 'This is HomePage component!',
-  },
-});
